Squashed 'libsrcs/libRocket/libRocket/' changes from 3a75ee7..5eacb8b
git-subtree-dir: libsrcs/libRocket/libRocket
git-subtree-split: 5eacb8b76443370143214f8b6e34625a940592cc

+++ Source/Core/FontDatabase.cpp
@@ -26,9 +26,9 @@
  */
 
 #include "precompiled.h"
-#include <Rocket/Core/FontDatabase.h>
+#include "../../Include/Rocket/Core/FontDatabase.h"
 #include "FontFamily.h"
-#include <Rocket/Core.h>
+#include "../../Include/Rocket/Core.h"
 #include <ft2build.h>
 #include FT_FREETYPE_H
 
@@ -288,7 +288,7 @@ void* FontDatabase::LoadFace(const String& file_name)
 	file_interface->Read(buffer, length, handle);
 	file_interface->Close(handle);
 
-	return LoadFace(buffer, length, file_name, true);
+	return LoadFace(buffer, (int)length, file_name, true);
 }
 
 // Loads a FreeType face from memory.
